removed requests dependencies and switched to authlib oauth2 authentication flow
