+++ fhir_kindling/fhir_server/auth.py
@@ -1,6 +1,7 @@
 import os
-import requests
-from requests.auth import HTTPBasicAuth
+from typing import Union, Tuple
+
+import httpx
 
 
 def load_environment_auth_vars() -> tuple:
@@ -19,7 +20,7 @@ def load_environment_auth_vars() -> tuple:
 
 
 def generate_auth(username: str = None, password: str = None, token: str = None,
-                  load_env: bool = False) -> requests.auth.AuthBase:
+                  load_env: bool = False) -> httpx.Auth:
     """Generate authentication for the request to be sent to server. Either based on a given bearer token or using basic
     auth with username and password.
 
@@ -48,7 +49,7 @@ def generate_auth(username: str = None, password: str = None, token: str = None,
         raise ValueError(f"Missing password for user: {username}")
 
     if username and password:
-        return HTTPBasicAuth(username=username, password=password)
+        return httpx.BasicAuth(username=username, password=password)
     elif token:
         return BearerAuth(token=token)
 
@@ -56,14 +57,49 @@ def generate_auth(username: str = None, password: str = None, token: str = None,
         raise ValueError("No authentication information given")
 
 
-class BearerAuth(requests.auth.AuthBase):
+def auth_info_from_env() -> Union[str, Tuple[str, str], Tuple[str, str, str]]:
+    # First try to load basic auth information
+    username = os.getenv("FHIR_USER")
+    # Static token auth
+    token = os.getenv("FHIR_TOKEN")
+    # oauth2/oidc authentication
+    client_id = os.getenv("CLIENT_ID")
+    client_secret = os.getenv("CLIENT_SECRET")
+    oidc_provider_url = os.getenv("OIDC_PROVIDER_URL")
+
+    if username and token:
+        raise EnvironmentError("Conflicting auth information, bother username and token present.")
+    if username and client_id:
+        raise EnvironmentError("Conflicting auth information, bother username and client id present")
+    if token and client_id:
+        raise EnvironmentError("Conflicting auth information, bother static token and client id present")
+
+    if username:
+        password = os.getenv("FHIR_PW")
+        if not password:
+            raise EnvironmentError(f"No password specified for user: {username}")
+        else:
+            print(f"Basic auth environment info found -> ({username}:******)")
+            return username, password
+    if token:
+        print("Found static auth token")
+        return token
+
+    if client_id and not client_secret:
+        raise EnvironmentError("Insufficient auth information, client id specified but no client secret found.")
+
+    if (client_id and client_secret) and not oidc_provider_url:
+        raise EnvironmentError("Insufficient auth information, client id and secret "
+                               "specified but no provider URL found")
+    if client_id and client_secret and oidc_provider_url:
+        print(f"Found OIDC auth configuration for client <{client_id}> with provider {oidc_provider_url}")
+        return client_id, client_secret, oidc_provider_url
+
+
+class BearerAuth(httpx.Auth):
     def __init__(self, token):
         self.token = token
 
-    def __call__(self, r):
-        r.headers["authorization"] = "Bearer " + self.token
-        return r
-
-    def __repr__(self):
-        return f"BearerAuth(token={self.token[:16]}...)"
-
+    def auth_flow(self, request):
+        request.headers['Authorization'] = f"Bearer {self.token}"
+        yield request
